function-app/requirements.txt
azure-functions
transformers 
PyMuPDF
openai
azure-storage-blob